Camera tracking integration ===========================
Automated image sequence offset calculation for image input node and image editor.

Calculate offset or sequence when loading image in node or editor and also
re-calculate offset when switching currently displaying image in editor or
source image in input node. This kind of recalculation is needed because offset
is stored in ImageUser structure so when you're switching from one sequence to
another it can be really confusing because offset from previous sequence would
be used for new sequence.

NOTE: Not sure this commit is fine to be included in tomato merge commit,
      prefer to test it further and commit as separate change after merge.